Refactors to only define SupportedCurrency type on front-end and typing it as a string on the backend.
