+++ packages/front-end/services/metrics.tsx
@@ -2,9 +2,9 @@ import { MetricInterface, MetricType } from "back-end/types/metric";
 import {
   GlobalPermission,
   ProjectScopedPermission,
-  SupportedCurrencies,
 } from "back-end/types/organization";
 import { PermissionFunctions } from "@/services/UserContext";
+import { SupportedCurrencies } from "@/pages/settings";
 
 const percentFormatter = new Intl.NumberFormat(undefined, {
   style: "percent",
